changes made as requested
