+++ budgetscreen.py
@@ -17,8 +17,6 @@
 # Import standard Python modules.
 import datetime
 import locale
-import ast, operator
-import re
 
 from gettext import gettext as _
 
@@ -29,6 +27,7 @@ from gi.repository import GObject
 from sugar3.graphics import style
 
 import colors
+from extractinput import evaluate
 
 # copied from finance.py to not create another module
 DAY = 0
@@ -244,8 +243,8 @@ class BudgetScreen(Gtk.VBox):
             try:
                 amount = abs(locale.atof(text))
             except ValueError :
-                if self._extract_value(text) is not None:
-                    amount = abs(locale.atof(self._extract_value(text)))
+                if evaluate(text) is not None:
+                    amount = abs(locale.atof(evaluate(text)))
                 else:
                     amount = 0.00
             if amount:
@@ -253,40 +252,3 @@ class BudgetScreen(Gtk.VBox):
         else:
             del self.activity.data['budgets'][category]
         self.queue_draw()
-
-    def _extract_value(self, value):
-        if isinstance(value, str):
-
-            binOps = {
-                ast.Add: operator.add,
-                ast.Sub: operator.sub,
-                ast.Mult: operator.mul,
-                ast.Div: operator.truediv,
-            }
-
-            try:
-                node = ast.parse(value, mode='eval')
-            except:
-                return None
-
-            def _eval(node):
-                if isinstance(node, ast.Expression):
-                    return _eval(node.body)
-                elif isinstance(node, ast.BinOp):
-                    return binOps[type(node.op)](_eval(node.left), _eval(node.right))
-                elif isinstance(node, ast.Num):
-                    return node.n
-                else:
-                    return None 
-                return _eval(node.body)
-
-            value = _eval(node) 
-            
-        decimals_found = re.findall("\d+\.\d+", str(value))
-        integers_found = re.findall("\d+", str(value))
-
-        if decimals_found != []:
-            return decimals_found[0]
-        elif integers_found != []:
-            return integers_found[0]
-        return None
