+++ registerscreen.py
@@ -20,8 +20,6 @@ import datetime
 import locale
 from gettext import gettext as _
 import logging
-import ast, operator
-import re
 
 from gi.repository import Gtk
 from gi.repository import GObject
@@ -31,6 +29,7 @@ from sugar3.graphics import style
 
 # Import activity module
 import colors
+from extractinput import evaluate
 
 REGISTER_HELP = _(
     'Welcome to Finance! This activity keeps track of income '
@@ -173,48 +172,11 @@ class RegisterScreen(Gtk.VBox):
         try:
             t['amount'] = abs(locale.atof(new_text))
         except ValueError :
-            if self._extract_value(new_text) is not None:
-                t['amount'] = abs(locale.atof(self._extract_value(new_text)))
+            if evaluate(new_text) is not None:
+                t['amount'] = abs(locale.atof(evaluate(new_text)))
             else:
-                t['amount'] = 0.00
+                return
         self.activity.update_summary()
-
-    def _extract_value(self, value):
-        if isinstance(value, str):
-
-            binOps = {
-                ast.Add: operator.add,
-                ast.Sub: operator.sub,
-                ast.Mult: operator.mul,
-                ast.Div: operator.truediv,
-            }
-
-            try:
-                node = ast.parse(value, mode='eval')
-            except:
-                return None
-            
-            def _eval(node):
-                if isinstance(node, ast.Expression):
-                    return _eval(node.body)
-                elif isinstance(node, ast.BinOp):
-                    return binOps[type(node.op)](_eval(node.left), _eval(node.right))
-                elif isinstance(node, ast.Num):
-                    return node.n
-                else:
-                    return None 
-                return _eval(node.body)
-
-            value = _eval(node) 
-            
-        decimals_found = re.findall("\d+\.\d+", str(value))
-        integers_found = re.findall("\d+", str(value))
-
-        if decimals_found != []:
-            return decimals_found[0]
-        elif integers_found != []:
-            return integers_found[0]
-        return None
 
     def date_render_cb(self, column, cell_renderer, model, iter, data):
         id = model.get_value(iter, 0)
